Modifies test tree; Fixes Rotation; Adds traversal property test
The test tree was incorrectly being generated with no parent
information set on the nodes, making certain tests impossible. THis
has been fixed.

The rotation functions are now hardened and ready to go. They have
been simplified to remove uneeded variables, ensure the parent
relationships are preserved.

The traversal property test ensures a key property of the rotations,
that the order of traversal is left untouched. With this we can say
for certain that these functions behave as expected.